organised the todopage by making a component
