+++ client/todo-app/package.json
@@ -18,6 +18,7 @@
     "axios": "^1.8.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.0",
     "tailwindcss": "^4.0.12"
   },
   "devDependencies": {
